All Dependent Data Files
DependentData2.zip is actually the output of GetEDA.py. Since we need to build mongodb to test the full function of GetEDA.py, I didn't do so in the testing phase. Considering FAA is using Oracle database, you may want to change the mongodb part to oracle. Thus, I just provide the output of GetEDA.py for IAH --> BOS case.